requirements-test.txt
black==19.10b0
flake8==3.8.3
PyJWT==1.7.1
pylint==2.5.3
tornado==6.0.4
tornadose==0.4.1
